requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
albumentations>=1.3.0
Pillow>=9.0.0
numpy>=1.21.0
tifffile>=2022.0.0
imagecodecs>=2022.0.0
